Refactored routing_import.py
