+++ quicklogic/utils/routing_import.py
@@ -3,7 +3,6 @@ import argparse
 import pickle
 import itertools
 import re
-from collections import defaultdict
 
 import lxml.etree as ET
 
@@ -13,16 +12,12 @@ import lib.rr_graph_xml.graph2 as rr_xml
 from lib import progressbar_utils
 
 from data_structs import *
-from utils import yield_muxes, fixup_pin_name
+from utils import fixup_pin_name
+
+from rr_utils import add_node, add_track, add_edge, connect
+from switchbox_model import SwitchboxModel
 
 # =============================================================================
-
-CLOCK_CELLS = [
-    "CLOCK",
-    "GMUX",
-    "QMUX",
-    "CAND"
-]
 
 def is_hop(connection):
     """
@@ -80,579 +75,6 @@ def is_local(connection):
     Returns true if a connection is local.
     """
     return connection.src.loc == connection.dst.loc
-
-
-# =============================================================================
-
-
-def add_track(graph, track, segment_id, node_timing=None):
-    """
-    Adds a track to the graph. Returns the node object representing the track
-    node.
-    """
-
-    if node_timing is None:
-        node_timing = rr.NodeTiming(r=0.0, c=0.0)
-
-    node_id = graph.add_track(track, segment_id, timing=node_timing)
-    node = graph.nodes[-1]
-    assert node.id == node_id
-
-    return node
-
-
-def add_node(graph, loc, direction, segment_id):
-    """
-    Adds a track of length 1 to the graph. Returns the node object
-    """
-
-    return add_track(
-        graph,
-        tracks.Track(
-            direction=direction,
-            x_low=loc.x,
-            x_high=loc.x,
-            y_low=loc.y,
-            y_high=loc.y,
-        ), segment_id
-    )
-
-
-# =============================================================================
-
-
-def node_joint_location(node_a, node_b):
-    """
-    Given two VPR nodes returns a location of the point where they touch each
-    other.
-    """
-
-    loc_a1 = Loc(node_a.loc.x_low, node_a.loc.y_low)
-    loc_a2 = Loc(node_a.loc.x_high, node_a.loc.y_high)
-
-    loc_b1 = Loc(node_b.loc.x_low, node_b.loc.y_low)
-    loc_b2 = Loc(node_b.loc.x_high, node_b.loc.y_high)
-
-    if loc_a1 == loc_b1:
-        return loc_a1
-    if loc_a1 == loc_b2:
-        return loc_a1
-    if loc_a2 == loc_b1:
-        return loc_a2
-    if loc_a2 == loc_b2:
-        return loc_a2
-
-    assert False, (node_a, node_b)
-
-
-def add_edge(
-        graph, src_node_id, dst_node_id, switch_id, meta_name=None,
-        meta_value=""
-):
-    """
-    Adds an edge to the routing graph. If the given switch corresponds to a
-    "pass" type switch then adds two edges going both ways.
-    """
-
-    # Sanity check
-    assert src_node_id != dst_node_id, \
-        (src_node_id, dst_node_id, switch_id, meta_name, meta_value)
-
-    # Connect src to dst
-    graph.add_edge(src_node_id, dst_node_id, switch_id, meta_name, meta_value)
-
-    # Check if the switch is of the "pass" type. If so then add an edge going
-    # in the opposite way.
-    switch = graph.switch_map[switch_id]
-    if switch.type in [rr.SwitchType.SHORT, rr.SwitchType.PASS_GATE]:
-
-        graph.add_edge(
-            dst_node_id, src_node_id, switch_id, meta_name, meta_value
-        )
-
-
-def connect(
-        graph,
-        src_node,
-        dst_node,
-        switch_id=None,
-        segment_id=None,
-        meta_name=None,
-        meta_value=""
-):
-    """
-    Connect two VPR nodes in a way that certain rules are obeyed.
-
-    The rules are:
-    - a CHANX cannot connect directly to a CHANY and vice versa,
-    - a CHANX cannot connect to an IPIN facing left or right,
-    - a CHANY cannot connect to an IPIN facting top or bottom,
-    - an OPIN facing left or right cannot connect to a CHANX,
-    - an OPIN facing top or bottom cannot connect to a CHANY
-
-    Whenever a rule is not met then the connection is made through a padding
-    node:
-
-    src -> [delayless] -> pad -> [desired switch] -> dst
-
-    Otherwise the connection is made directly
-
-    src -> [desired switch] -> dst
-
-    The influence of whether the rules are obeyed or not on the actual VPR
-    behavior is unclear.
-    """
-
-    # Use the default delayless switch if none is given
-    if switch_id is None:
-        switch_id = graph.get_delayless_switch_id()
-
-    # Determine which segment to use if none given
-    if segment_id is None:
-        # If the source is IPIN/OPIN then use the same segment as used by
-        # the destination.
-        # If the destination is IPIN/OPIN then do the opposite.
-        # Finally if both are CHANX/CHANY then use the source's segment.
-
-        if src_node.type in [rr.NodeType.IPIN, rr.NodeType.OPIN]:
-            segment_id = dst_node.segment.segment_id
-        elif dst_node.type in [rr.NodeType.IPIN, rr.NodeType.OPIN]:
-            segment_id = src_node.segment.segment_id
-        else:
-            segment_id = src_node.segment.segment_id
-
-    # CHANX to CHANY or vice-versa
-    if src_node.type == rr.NodeType.CHANX and dst_node.type == rr.NodeType.CHANY or \
-       src_node.type == rr.NodeType.CHANY and dst_node.type == rr.NodeType.CHANX:
-
-        # Check loc
-        node_joint_location(src_node, dst_node)
-
-        # Connect directly
-        add_edge(
-            graph, src_node.id, dst_node.id, switch_id, meta_name, meta_value
-        )
-
-    # CHANX to CHANX or CHANY to CHANY
-    elif src_node.type == rr.NodeType.CHANX and dst_node.type == rr.NodeType.CHANX or \
-         src_node.type == rr.NodeType.CHANY and dst_node.type == rr.NodeType.CHANY:
-
-        loc = node_joint_location(src_node, dst_node)
-        direction = "X" if src_node.type == rr.NodeType.CHANY else "Y"
-
-        # Padding node
-        pad_node = add_node(graph, loc, direction, segment_id)
-
-        # Connect through the padding node
-        add_edge(
-            graph, src_node.id, pad_node.id, graph.get_delayless_switch_id()
-        )
-
-        add_edge(
-            graph, pad_node.id, dst_node.id, switch_id, meta_name, meta_value
-        )
-
-    # OPIN to CHANX/CHANY
-    elif src_node.type == rr.NodeType.OPIN and dst_node.type in \
-        [rr.NodeType.CHANX, rr.NodeType.CHANY]:
-
-        # All OPINs go right (towards +X)
-        assert src_node.loc.side == tracks.Direction.RIGHT, src_node
-
-        # Connected to CHANX
-        if dst_node.type == rr.NodeType.CHANX:
-
-            loc = node_joint_location(src_node, dst_node)
-
-            # Padding node
-            pad_node = add_node(graph, loc, "Y", segment_id)
-
-            # Connect through the padding node
-            add_edge(
-                graph, src_node.id, pad_node.id,
-                graph.get_delayless_switch_id()
-            )
-
-            add_edge(
-                graph, pad_node.id, dst_node.id, switch_id, meta_name,
-                meta_value
-            )
-
-        # Connected to CHANY
-        elif dst_node.type == rr.NodeType.CHANY:
-
-            # Directly
-            add_edge(
-                graph, src_node.id, dst_node.id, switch_id, meta_name,
-                meta_value
-            )
-
-        # Should not happen
-        else:
-            assert False, dst_node
-
-    # CHANX/CHANY to IPIN
-    elif dst_node.type == rr.NodeType.IPIN and src_node.type in \
-        [rr.NodeType.CHANX, rr.NodeType.CHANY]:
-
-        # All IPINs go top (toward +Y)
-        assert dst_node.loc.side == tracks.Direction.TOP, dst_node
-
-        # Connected to CHANY
-        if src_node.type == rr.NodeType.CHANY:
-
-            loc = node_joint_location(src_node, dst_node)
-
-            # Padding node
-            pad_node = add_node(graph, loc, "X", segment_id)
-
-            # Connect through the padding node
-            add_edge(
-                graph, src_node.id, pad_node.id,
-                graph.get_delayless_switch_id()
-            )
-
-            add_edge(
-                graph, pad_node.id, dst_node.id, switch_id, meta_name,
-                meta_value
-            )
-
-        # Connected to CHANX
-        elif src_node.type == rr.NodeType.CHANX:
-
-            # Directly
-            add_edge(
-                graph, src_node.id, dst_node.id, switch_id, meta_name,
-                meta_value
-            )
-
-        # Should not happen
-        else:
-            assert False, dst_node
-
-    # An unhandled case
-    else:
-        assert False, (src_node, dst_node)
-
-
-# =============================================================================
-
-
-class SwitchboxModel(object):
-    """
-    Represents a model of connectivity of a concrete instance of a switchbox.
-    """
-
-    def __init__(self, graph, loc, phy_loc, switchbox):
-        self.graph = graph
-        self.loc = loc
-        self.phy_loc = phy_loc
-        self.switchbox = switchbox
-
-        self.mux_input_to_node = {}
-        self.mux_output_to_node = {}
-
-        self.input_to_node = {}
-
-        self._build()
-
-    @staticmethod
-    def get_chan_dirs_for_stage(stage):
-        """
-        Returns channel directions for inputs and outputs of a stage.
-        """
-
-        if stage.type == "HIGHWAY":
-            return "Y", "X"
-
-        elif stage.type == "STREET":
-            dir_inp = "Y" if (stage.id % 2) else "X"
-            dir_out = "X" if (stage.id % 2) else "Y"
-            return dir_inp, dir_out
-
-        else:
-            assert False, stage.type
-
-    def _create_muxes(self):
-        """
-        Creates nodes for muxes and internal edges within them. Annotates the
-        internal edges with fasm data.
-
-        Builds maps of muxs' inputs and outpus to VPR nodes.
-        """
-
-        # Build mux driver timing map. Assign each mux output its timing data
-        driver_timing = {}
-        for connection in self.switchbox.connections:
-            src = connection.src
-            dst = connection.dst
-
-            stage = self.switchbox.stages[src.stage_id]
-            switch = stage.switches[src.switch_id]
-            mux = switch.muxes[src.mux_id]
-            pin = mux.inputs[src.pin_id]
-
-            if pin.id not in mux.timing:
-                continue
-
-            timing = mux.timing[pin.id].driver
-
-            key = (src.stage_id, src.switch_id, src.mux_id)
-            if key in driver_timing:
-                assert driver_timing[key] == timing, \
-                    (self.loc, key, driver_timing[key], timing)
-            else:
-                driver_timing[key] = timing
-
-        # Create muxes
-        segment_id = self.graph.get_segment_id_from_name("sbox")
-
-        for stage, switch, mux in yield_muxes(self.switchbox):
-            dir_inp, dir_out = self.get_chan_dirs_for_stage(stage)
-
-            # Output node
-            key = (stage.id, switch.id, mux.id)
-            assert key not in self.mux_output_to_node
-
-            out_node = add_node(self.graph, self.loc, dir_out, segment_id)
-            self.mux_output_to_node[key] = out_node
-
-            # Intermediate output node
-            int_node = add_node(self.graph, self.loc, dir_out, segment_id)
-
-            # Get switch id for the switch assigned to the driver. If
-            # there is none then use the delayless switch. Probably the
-            # driver is connected to a const.
-            if key in driver_timing:
-                switch_id = self.graph.get_switch_id(
-                    driver_timing[key].vpr_switch
-                )
-            else:
-                switch_id = self.graph.get_delayless_switch_id()
-
-            # Output driver edge
-            connect(
-                self.graph,
-                int_node,
-                out_node,
-                switch_id=switch_id,
-                segment_id=segment_id,
-            )
-
-            # Input nodes + mux edges
-            for pin in mux.inputs.values():
-
-                key = (stage.id, switch.id, mux.id, pin.id)
-                assert key not in self.mux_input_to_node
-
-                # Input node
-                inp_node = add_node(self.graph, self.loc, dir_inp, segment_id)
-                self.mux_input_to_node[key] = inp_node
-
-                # Get mux metadata
-                metadata = self._get_metadata_for_mux(
-                    stage, switch, mux, pin.id
-                )
-
-                if len(metadata):
-                    meta_name = "fasm_features"
-                    meta_value = "\n".join(metadata)
-                else:
-                    meta_name = None
-                    meta_value = ""
-
-                # Get switch id for the switch assigned to the mux edge. If
-                # there is none then use the delayless switch. Probably the
-                # edge is connected to a const.
-                if pin.id in mux.timing:
-                    switch_id = self.graph.get_switch_id(
-                        mux.timing[pin.id].sink.vpr_switch
-                    )
-                else:
-                    switch_id = self.graph.get_delayless_switch_id()
-
-                # Mux switch with appropriate timing and fasm metadata
-                connect(
-                    self.graph,
-                    inp_node,
-                    int_node,
-                    switch_id=switch_id,
-                    segment_id=segment_id,
-                    meta_name=meta_name,
-                    meta_value=meta_value,
-                )
-
-    def _connect_muxes(self):
-        """
-        Creates VPR edges that connects muxes within the switchbox.
-        """
-
-        segment_id = self.graph.get_segment_id_from_name("sbox")
-        switch_id = self.graph.get_switch_id("short")
-
-        # Add internal connections between muxes.
-        for connection in self.switchbox.connections:
-            src = connection.src
-            dst = connection.dst
-
-            # Check
-            assert src.pin_id == 0, src
-            assert src.pin_direction == PinDirection.OUTPUT, src
-
-            # Get the input node
-            key = (dst.stage_id, dst.switch_id, dst.mux_id, dst.pin_id)
-            dst_node = self.mux_input_to_node[key]
-
-            # Get the output node
-            key = (src.stage_id, src.switch_id, src.mux_id)
-            src_node = self.mux_output_to_node[key]
-
-            # Connect
-            connect(
-                self.graph,
-                src_node,
-                dst_node,
-                switch_id=switch_id,
-                segment_id=segment_id
-            )
-
-    def _create_input_drivers(self):
-        """
-        Creates VPR nodes and edges that model input connectivity of the
-        switchbox.
-        """
-
-        # Create a driver map containing all mux pin locations that are
-        # connected to a driver. The map is indexed by (pin_name, vpr_switch)
-        # and groups togeather inputs that should be driver by a specific
-        # switch due to the timing model.
-        driver_map = defaultdict(lambda: [])
-
-        for pin in self.switchbox.inputs.values():
-            for loc in pin.locs:
-
-                stage = self.switchbox.stages[loc.stage_id]
-                switch = stage.switches[loc.switch_id]
-                mux = switch.muxes[loc.mux_id]
-                pin = mux.inputs[loc.pin_id]
-
-                if pin.id not in mux.timing:
-                    vpr_switch = None
-                else:
-                    vpr_switch = mux.timing[pin.id].driver.vpr_switch
-
-                key = (pin.name, vpr_switch)
-                driver_map[key].append(loc)
-
-        # Create input nodes for each input pin
-        segment_id = self.graph.get_segment_id_from_name("sbox")
-
-        for pin in self.switchbox.inputs.values():
-
-            node = add_node(self.graph, self.loc, "Y", segment_id)
-
-            assert pin.name not in self.input_to_node, pin.name
-            self.input_to_node[pin.name] = node
-
-        # Create driver nodes, connect everything
-        for (pin_name, vpr_switch), locs in driver_map.items():
-
-            # Create the driver node
-            drv_node = add_node(self.graph, self.loc, "X", segment_id)
-
-            # Connect input node to the driver node. Use the switch with timing.
-            inp_node = self.input_to_node[pin_name]
-
-            # Get switch id for the switch assigned to the driver. If
-            # there is none then use the delayless switch. Probably the
-            # driver is connected to a const.
-            if vpr_switch is not None:
-                switch_id = self.graph.get_switch_id(vpr_switch)
-            else:
-                switch_id = self.graph.get_delayless_switch_id()
-
-            # Connect
-            connect(
-                self.graph,
-                inp_node,
-                drv_node,
-                switch_id=switch_id,
-                segment_id=segment_id
-            )
-
-            # Now connect the driver node with its loads
-            switch_id = self.graph.get_switch_id("short")
-            for loc in locs:
-
-                key = (loc.stage_id, loc.switch_id, loc.mux_id, loc.pin_id)
-                dst_node = self.mux_input_to_node[key]
-
-                connect(
-                    self.graph,
-                    drv_node,
-                    dst_node,
-                    switch_id=switch_id,
-                    segment_id=segment_id
-                )
-
-    def _build(self):
-        """
-        Build the switchbox model
-        """
-
-        # Create and connect muxes
-        self._create_muxes()
-        self._connect_muxes()
-
-        # Create and connect input drivers models
-        self._create_input_drivers()
-
-    def _get_metadata_for_mux(self, stage, switch, mux, src_pin_id):
-        """
-        Formats fasm features for the given edge representin a switchbox mux.
-        Returns a list of fasm features.
-        """
-        metadata = []
-
-        # Format prefix
-        prefix = "X{}Y{}.ROUTING".format(self.phy_loc.x, self.phy_loc.y)
-
-        # A mux in the HIGHWAY stage
-        if stage.type == "HIGHWAY":
-            feature = "I_highway.IM{}.I_pg{}".format(switch.id, src_pin_id)
-
-        # A mux in the STREET stage
-        elif stage.type == "STREET":
-            feature = "I_street.Isb{}{}.I_M{}.I_pg{}".format(
-                stage.id + 1, switch.id + 1, mux.id, src_pin_id
-            )
-
-        else:
-            assert False, stage
-
-        metadata.append(".".join([prefix, feature]))
-        return metadata
-
-    def get_input_node(self, pin_name):
-        """
-        Returns a VPR node associated with the given input of the switchbox
-        """
-        return self.input_to_node[pin_name]
-
-    def get_output_node(self, pin_name):
-        """
-        Returns a VPR node associated with the given output of the switchbox
-        """
-
-        # Get the output pin
-        pin = self.switchbox.outputs[pin_name]
-
-        assert len(pin.locs) == 1
-        loc = pin.locs[0]
-
-        # Return its node
-        key = (loc.stage_id, loc.switch_id, loc.mux_id)
-        return self.mux_output_to_node[key]
 
 
 # =============================================================================
